pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anny"
version = "0.1.0"
description = "A modern Python web application built with FastAPI"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "google-auth>=2.0.0",
    "google-analytics-data>=0.18.0",
    "google-api-python-client>=2.100.0",
    "fastmcp>=2.0.0,<3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pylint.format]
max-line-length = 100

[tool.pylint."messages control"]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
    "import-error",
    "too-many-arguments",
    "too-many-positional-arguments",
]
